[atlas-graddle-plugin] complete bundle proguard

+++ atlas-gradle-plugin/atlas-plugin/src/main/java/com/taobao/android/builder/tools/proguard/BundleProguarder.java
@@ -225,8 +225,9 @@ import com.taobao.android.builder.dependency.model.AwbBundle;
 import com.taobao.android.builder.tools.ReflectUtils;
 import com.taobao.android.builder.tools.cache.FileCache.SimpleFileCache;
 import com.taobao.android.builder.tools.cache.FileLockUtils;
-import com.taobao.android.builder.tools.proguard.dto.Input;
-import com.taobao.android.builder.tools.proguard.dto.Result;
+import com.taobao.android.builder.tools.proguard.domain.Input;
+import com.taobao.android.builder.tools.proguard.domain.LibClassRefVisitor;
+import com.taobao.android.builder.tools.proguard.domain.Result;
 import com.taobao.android.builder.tools.zip.ZipUtils;
 import org.apache.commons.io.FileUtils;
 import org.gradle.api.GradleException;
@@ -248,7 +249,7 @@ import static proguard.AtlasProguardConstants.INOUT_CFG;
  */
 public class BundleProguarder {
 
-    public static final String CACHE_TYPE = "proguard-bundles-v1.1";
+    public static final String CACHE_TYPE = "proguard-bundles-v1.5";
 
     private static Logger logger = LoggerFactory.getLogger(BundleProguarder.class);
 
@@ -282,8 +283,18 @@ public class BundleProguarder {
                             }
                         }
                     }
+
+                    //缓存keep.json
                     FileUtils.copyFileToDirectory(input.getAwbBundles().get(0).getAwbBundle().getKeepProguardFile(),
                                                   result.cacheDir);
+
+                    File proguardCfg = new File(null != input.proguardOutputDir ? input.proguardOutputDir
+                                                    : appVariantContext
+                                                        .getAwbProguardDir(input.getAwbBundles().get(0).getAwbBundle()),
+                                                "proguard.cfg");
+                    if (proguardCfg.exists()) {
+                        FileUtils.copyFileToDirectory(proguardCfg, result.cacheDir);
+                    }
 
                 } catch (Exception e) {
                     throw new GradleException(e.getMessage());
@@ -315,6 +326,8 @@ public class BundleProguarder {
         logger.warn("bundle proguard for " + name + " hit  cache ");
 
         File keepJson = new File(cacheDir, "keep.json");
+        File proguardCfg = new File(cacheDir, "proguard.cfg");
+
         if (!keepJson.exists()) {
             logger.error("bundle proguard for " + name + " missing keep.json  ");
             FileUtils.deleteDirectory(cacheDir);
@@ -328,12 +341,27 @@ public class BundleProguarder {
                 }
             }
 
+            File awbProguardDir = input.proguardOutputDir;
+            FileUtils.copyFileToDirectory(keepJson, awbProguardDir);
+            if (proguardCfg.exists()) {
+                FileUtils.copyFileToDirectory(proguardCfg,
+                                              awbProguardDir);
+            }
+
             result.success = true;
             return result;
         }
 
         Map<AwbTransform, List<File>> transformListMap = new HashMap<>();
         for (AwbTransform awbTransform : input.getAwbBundles()) {
+
+            File awbProguardDir = appVariantContext.getAwbProguardDir(awbTransform.getAwbBundle());
+            FileUtils.copyFileToDirectory(keepJson, awbProguardDir);
+            if (proguardCfg.exists()) {
+                FileUtils.copyFileToDirectory(proguardCfg,
+                                              awbProguardDir);
+            }
+
             List<File> inputFiles = new ArrayList<>();
             transformListMap.put(awbTransform, inputFiles);
 
@@ -360,7 +388,7 @@ public class BundleProguarder {
             }
         }
 
-        for (AwbTransform awbTransform : input.getAwbBundles()){
+        for (AwbTransform awbTransform : input.getAwbBundles()) {
             awbTransform.setInputFiles(transformListMap.get(awbTransform));
             awbTransform.setInputDir(null);
             awbTransform.getInputLibraries().clear();
@@ -400,19 +428,19 @@ public class BundleProguarder {
         ClassPool classPool = (ClassPool)ReflectUtils.getField(proGuard, "programClassPool");
         //ClassPool libraryClassPool = (ClassPool)ReflectUtils.getField(proGuard, "libraryClassPool");
 
-        LibClassRefVisitor classRefPrinter = new LibClassRefVisitor(input.getDefaultLibraryClasses(),classPool);
+        LibClassRefVisitor classRefPrinter = new LibClassRefVisitor(input.getDefaultLibraryClasses(), classPool);
         classPool.classesAccept(classRefPrinter);
 
         //Fileoutputs
         for (AwbTransform awbTransform : input.getAwbBundles()) {
-
-            AwbBundle awbBundle = awbTransform.getAwbBundle();
-            File fileOut = new File(appVariantContext.getAwbProguardDir(awbBundle), "keep.json");
-            fileOut.delete();
-            fileOut.getParentFile().mkdirs();
-            FileUtils.write(fileOut, JSON.toJSONString(classRefPrinter.getRefClazzMap()));
-            awbBundle.setKeepProguardFile(fileOut);
-            //File json = new File(appVariantContext.getAwbProguardDir(awbBundle), "mainDexClazzRef.json");
+            if (!awbTransform.getAwbBundle().isMainBundle()) {
+                AwbBundle awbBundle = awbTransform.getAwbBundle();
+                File fileOut = new File(appVariantContext.getAwbProguardDir(awbBundle), "keep.json");
+                fileOut.delete();
+                fileOut.getParentFile().mkdirs();
+                FileUtils.write(fileOut, JSON.toJSONString(classRefPrinter.getRefClazzMap()));
+                awbBundle.setKeepProguardFile(fileOut);
+            }
         }
 
     }
@@ -459,8 +487,6 @@ public class BundleProguarder {
 
         File inoutConfigs = new File(proguardDir, INOUT_CFG);
         List<String> configs = new ArrayList<>();
-
-
 
         for (AwbTransform awbTransform : input.getAwbBundles()) {
